refactor: rename HmacAuthenticationValidation to HeaderValidation
