+++ src/Sinch/Conversation/Webhooks/HeaderValidation.cs
@@ -9,7 +9,7 @@ namespace Sinch.Conversation.Webhooks;
 /// <summary>
 ///     Validates HMAC authentication headers for Conversation webhook requests.
 /// </summary>
-internal sealed class HmacAuthenticationValidation
+internal sealed class HeaderValidation
 {
     private const string TimestampHeader = "x-sinch-webhook-signature-timestamp";
     private const string NonceHeader = "x-sinch-webhook-signature-nonce";
